requirements.txt
python-jobspy==1.1.80
pandas==2.2.3
python-dotenv==1.0.1
requests==2.32.3
